remake FE folders structure, delete timetable, prepare user profile
