+++ src/Components/form/form.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Login} from './login';
 import {Register} from './register';
-import {User} from '../Service/user-service';
+import {User} from '../../Service/user-service';
 
 export class Form extends React.Component {
